Incremental: wip

+++ packages/incremental/computations/mixins/cacheable.ts
@@ -1,10 +1,9 @@
+import * as lmdb from "lmdb";
 import fsextra from "fs-extra";
+import path from "node:path";
+import { inspect } from "node:util";
 import { never } from "../../../util/miscellaneous";
-import {
-  AnyComputationDescription,
-  ResultTypeOfComputation,
-  type ComputationDescription,
-} from "../../incremental-lib";
+import { ResultTypeOfComputation } from "../../incremental-lib";
 import { type ValueDefinition, HashMap, objValue } from "../../utils/hash-map";
 import {
   ComputationResult,
@@ -17,14 +16,13 @@ import {
   RawComputationContext,
   RawComputationExec,
 } from "../raw";
+import type { ComputationDescription } from "../description";
 import { DependentContext, MaybeDependentComputation } from "./dependent";
 import { MaybeParentComputation, MaybeParentContext } from "./parent";
 import { CtxWithFS } from "../file-system/file-system";
-import { SerializationDB } from "../../../util/serialization";
+import { MissingConstructorSerializerError } from "../../../util/serialization";
 import { SubscribableComputation } from "./subscribable";
 import { ChildComputation } from "./child";
-
-export type CacheableTag = string;
 
 export type CachedGet = {
   readonly kind: "get";
@@ -47,22 +45,63 @@ class CacheEntry<Res> {
   constructor(
     readonly value: Res,
     readonly deps: readonly CachedDep[],
-    readonly version: number
+    readonly version: number,
+    readonly useDeps: boolean,
+    readonly alive: boolean // If true, keep it. Otherwise, GC it
   ) {}
 }
 
-// TODO? https://github.com/parcel-bundler/parcel/blob/v2/packages/core/utils/src/stream.js
-// https://github.com/parcel-bundler/parcel/blob/v2/packages/core/cache/src/FSCache.js
+type CachedDepInDisk =
+  | {
+      readonly kind: "get";
+      readonly desc: ComputationDescription<any>;
+    }
+  | {
+      readonly kind: "compute";
+      readonly desc: ComputationDescription<any>;
+    };
+
+type CacheEntryInDisk = {
+  readonly desc: ComputationDescription<any>;
+  readonly value: unknown;
+  readonly deps: readonly CachedDepInDisk[];
+  readonly useDeps: boolean;
+};
+
+export type CacheSaveOpts = {
+  readonly garbageCollect: boolean;
+};
+
+type DB_Val = Readonly<CacheEntryInDisk>[];
 
 export class CacheDB {
-  private locked = false;
-  private readonly map: HashMap<ComputationDescription<any>, CacheEntry<any>> =
-    new HashMap(defaultValDef);
+  private static DB_VERSION = 1;
+  private readonly dir: string;
+  private readonly logFile: string;
 
-  constructor(
-    private readonly dir: string,
-    private readonly serializationDB: SerializationDB
-  ) {}
+  private readonly map: HashMap<
+    ComputationDescription<any>,
+    CacheEntry<any> | null // Null for deleted entries
+  > = new HashMap(defaultValDef);
+
+  private locked = false;
+  private lastLog: Promise<void>;
+
+  constructor(dir: string) {
+    this.dir =
+      path.resolve(dir) + path.sep + `quase_incremental_v${CacheDB.DB_VERSION}`;
+    this.logFile = this.dir + path.sep + `log${Date.now()}.txt`;
+    this.lastLog = Promise.resolve();
+  }
+
+  _log(message: string, value: unknown = null) {
+    this.lastLog = this.lastLog.then(() => {
+      return fsextra.appendFile(
+        this.logFile,
+        `\n\n${message}\n${inspect(value)}`
+      );
+    });
+  }
 
   lock() {
     this.locked = true;
@@ -71,48 +110,209 @@ export class CacheDB {
   get<C extends AnyRawComputation>(desc: ComputationDescription<C>) {
     return this.map.get(desc) as
       | CacheEntry<ResultTypeOfComputation<C>>
+      | null
       | undefined;
   }
 
   set<C extends AnyRawComputation>(
     desc: ComputationDescription<C>,
-    entry: CacheEntry<ResultTypeOfComputation<C>> | null
+    entry: CacheEntry<ResultTypeOfComputation<C>>
   ) {
     if (this.locked) {
       return;
     }
-    if (entry != null) {
-      this.map.set(desc, entry);
-    }
+    this.map.set(desc, entry);
   }
 
   delete(desc: ComputationDescription<any>) {
     if (this.locked) {
       return;
     }
-    this.map.delete(desc);
+    this.map.set(desc, null);
   }
 
-  async save() {
-    // await fsextra.ensureDir(this.dir);
+  // When loading, we assign -1 to the versions to avoid confusion
+  // between versions cached in disk (from a previous session)
+  // and versions created at runtime in this session
+  async load() {
+    const cacheDir = this.dir;
+    const db = lmdb.open<DB_Val, string>({
+      path: cacheDir,
+    });
 
-    // TODO
-    console.log("========SAVING CACHE==========");
-    for (const [desc, entry] of this.map) {
-      console.log("========CACHE==========");
-      console.log({
-        ...desc,
-      });
-      console.log(entry);
+    this._log("=== LOADING CACHE ===");
+
+    // TODO loading everything to memory, is it ok? yes, big blobs should be handled separately
+    for (const key of db.getKeys()) {
+      try {
+        const dbValue = db.get(key) ?? [];
+        for (const entry of dbValue) {
+          this.set(entry.desc, this.unpackEntry(entry));
+          this._log("LOADED ENTRY", entry);
+        }
+      } catch (err) {
+        this._log(
+          "ERROR",
+          this.addError(
+            new Error(`Error loading entries with db key ${inspect(key)}`, {
+              cause: err,
+            })
+          )
+        );
+      }
     }
 
-    // TODO garbage collect old cache entries
+    await db.close();
+
+    this._log("=== CACHE LOADED ===");
+
+    this.printErrors();
+
+    await this.lastLog;
   }
 
-  async load() {
-    // TODO
-    // TODO make versions negative to avoid confusion between versions cached in disk (from a previous session),
-    // and versions created at runtime in this session
+  // When saving, we check that when a computation has a dependency, that version of the dependency is stored (see packEntry)
+  // This allows us to have a normalized cache and not store versions in disk (which in turn avoids future confusions)
+  // When loading, we assign -1 to all of them (see load())
+  // (e.g., it is possible that a dependency version is no longer available
+  // if the user interrupted the incremental library in the middle of something running)
+  async save(opts: CacheSaveOpts) {
+    const cacheDir = this.dir;
+    const db = lmdb.open<DB_Val, string>({
+      path: cacheDir,
+    });
+
+    this._log("=== SAVING CACHE ===");
+
+    for (const [desc, entry] of this.map) {
+      const key = desc.key().slice(0, 1978);
+      const packedEntry = entry ? this.packEntry(desc, entry, opts) : null;
+      try {
+        await db.transaction(async () => {
+          const entries = db.get(key) ?? [];
+
+          for (let i = 0; i < entries.length; i++) {
+            if (entries[i].desc.equal(desc)) {
+              if (packedEntry) {
+                entries[i] = packedEntry;
+              } else {
+                entries.splice(i, 1);
+              }
+              await db.put(key, entries);
+              return;
+            }
+          }
+
+          if (packedEntry) {
+            entries.push(packedEntry);
+            await db.put(key, entries);
+          }
+        });
+
+        if (packedEntry) {
+          this._log("SAVED ENTRY", packedEntry);
+        } else {
+          this._log("GC ENTRY", desc);
+        }
+      } catch (err) {
+        this._log(
+          "ERROR",
+          this.addError(
+            new Error(`Error saving entry with description ${inspect(desc)}`, {
+              cause: err,
+            })
+          )
+        );
+      }
+    }
+
+    await db.close();
+
+    this._log("=== SAVED CACHE ===");
+
+    this.printErrors();
+
+    await this.lastLog;
+  }
+
+  private errors: unknown[] = [];
+  private missingSerializers: Set<string> = new Set();
+
+  private addError(error: unknown) {
+    if (error instanceof MissingConstructorSerializerError) {
+      this.missingSerializers.add(error.constructorName);
+    }
+    if (
+      error instanceof Error &&
+      error.cause instanceof MissingConstructorSerializerError
+    ) {
+      this.missingSerializers.add(error.cause.constructorName);
+    }
+    this.errors.push(error);
+    return error;
+  }
+
+  private printErrors() {
+    for (const error of this.errors) {
+      console.error("Error when saving cache:", error);
+    }
+
+    if (this.missingSerializers.size) {
+      console.error("Missing serializers for:", ...this.missingSerializers);
+    }
+  }
+
+  private unpackEntry(entry: CacheEntryInDisk) {
+    const { value, deps, useDeps } = entry;
+    return new CacheEntry(
+      value,
+      deps.map(d => {
+        switch (d.kind) {
+          case "get":
+            return {
+              kind: "get",
+              desc: d.desc,
+              version: -1,
+            };
+          case "compute":
+            return {
+              kind: "compute",
+              desc: d.desc,
+            };
+          default:
+            never(d);
+        }
+      }),
+      -1,
+      useDeps,
+      false // When unpacking from disk, set this flag to false
+    );
+  }
+
+  private packEntry<Res>(
+    desc: ComputationDescription<any>,
+    entry: CacheEntry<Res>,
+    opts: CacheSaveOpts
+  ): CacheEntryInDisk | null {
+    if (opts.garbageCollect && !entry.alive) {
+      return null;
+    }
+    const useDeps =
+      entry.useDeps &&
+      entry.deps.every(d =>
+        d.kind === "get" ? this.get(d.desc)?.version === d.version : true
+      );
+    return {
+      desc,
+      value: entry.value,
+      deps: useDeps
+        ? entry.deps.map(d => ({
+            kind: d.kind,
+            desc: d.desc,
+          }))
+        : [],
+      useDeps,
+    };
   }
 }
 
@@ -147,12 +347,11 @@ export class CacheableComputationMixin<
     this.firstExec = true;
   }
 
-  finishRoutine({
-    result,
-    version,
-  }: VersionedComputationResult<
-    ResultTypeOfComputation<C>
-  >): VersionedComputationResult<ResultTypeOfComputation<C>> {
+  finishRoutine(
+    original: VersionedComputationResult<ResultTypeOfComputation<C>>,
+    storeDeps: boolean // Should be true if CacheableComputationMixin#exec is going to be used!
+  ): VersionedComputationResult<ResultTypeOfComputation<C>> {
+    const { result, version } = original;
     if (this.isCacheable) {
       const { firstExec } = this;
       this.firstExec = false;
@@ -160,11 +359,11 @@ export class CacheableComputationMixin<
       if (result.ok) {
         const calls: CachedDep[] = [];
 
-        if (this.source.dependentMixin) {
+        if (storeDeps && this.source.dependentMixin) {
           const getCalls = this.source.dependentMixin.getAllGetCalls();
           if (!getCalls) {
             this.db.delete(this.desc);
-            return { result, version };
+            return original;
           }
           for (const dep of getCalls) {
             calls.push({
@@ -175,7 +374,7 @@ export class CacheableComputationMixin<
           }
         }
 
-        if (this.source.parentMixin) {
+        if (storeDeps && this.source.parentMixin) {
           for (const child of this.source.parentMixin.getChildren()) {
             calls.push({
               kind: "compute",
@@ -184,10 +383,9 @@ export class CacheableComputationMixin<
           }
         }
 
-        let newEntry = new CacheEntry(result.value, calls, version);
-
+        let currentEntry;
         if (firstExec) {
-          const currentEntry = this.db.get(this.desc);
+          currentEntry = this.db.get(this.desc);
           if (currentEntry) {
             const cached = currentEntry.value;
             if (
@@ -195,21 +393,38 @@ export class CacheableComputationMixin<
               this.source.responseEqual(cached, result.value)
             ) {
               // If the final value is the same, keep the cached version number
-              newEntry = new CacheEntry(cached, calls, currentEntry.version);
+              const entry = new CacheEntry(
+                cached,
+                calls,
+                currentEntry.version,
+                storeDeps,
+                true
+              );
+              this.db.set(this.desc, entry);
+              this.db._log("REUSING", { desc: this.desc, entry });
+              return {
+                result: ok(cached),
+                version: currentEntry.version,
+              };
             }
           }
         }
 
-        this.db.set(this.desc, newEntry);
-        return {
-          result: ok(newEntry.value),
-          version: newEntry.version,
-        };
+        const entry = new CacheEntry(
+          result.value,
+          calls,
+          version,
+          storeDeps,
+          true
+        );
+        this.db.set(this.desc, entry);
+        this.db._log("NOT REUSING", { desc: this.desc, entry, currentEntry });
       } else {
         this.db.delete(this.desc);
+        this.db._log("DELETED", { desc: this.desc });
       }
     }
-    return { result, version };
+    return original;
   }
 
   invalidateRoutine() {
@@ -236,7 +451,9 @@ export class CacheableComputationMixin<
   ): Promise<ComputationResult<ResultTypeOfComputation<C>>> {
     if (this.isCacheable && this.firstExec) {
       const currentEntry = this.db.get(this.desc);
-      if (currentEntry) {
+      // If currentEntry.useDeps is false, it means the cache does not have the version of the dependency we need or that the computation depends on more than just the "ctx" calls
+      // So, it is not worth to run this, just execute the computation again and rely on "finishRoutine"
+      if (currentEntry && currentEntry.useDeps) {
         const cached = currentEntry.value;
         try {
           const jobs = [];
